chore(nvim): clean lazy configs
