--- .config/nvim/lua/senyc/home.lua
@@ -1,122 +1,0 @@
-local if_nil = vim.F.if_nil
-local M = {}
-
-local default_terminal = {
-  type = 'terminal',
-  command = nil,
-  width = 69,
-  height = 8,
-  opts = {
-    redraw = true,
-    window_config = {},
-  },
-}
-
-local default_header = {
-  type = 'text',
-  val = {
-    [[=================     ===============     ===============   ========  ========]],
-    [[\\ . . . . . . .\\   //. . . . . . .\\   //. . . . . . .\\  \\. . .\\// . . //]],
-    [[||. . ._____. . .|| ||. . ._____. . .|| ||. . ._____. . .|| || . . .\/ . . .||]],
-    [[|| . .||   ||. . || || . .||   ||. . || || . .||   ||. . || ||. . . . . . . ||]],
-    [[||. . ||   || . .|| ||. . ||   || . .|| ||. . ||   || . .|| || . | . . . . .||]],
-    [[|| . .||   ||. _-|| ||-_ .||   ||. . || || . .||   ||. _-|| ||-_.|\ . . . . ||]],
-    [[||. . ||   ||-'  || ||  `-||   || . .|| ||. . ||   ||-'  || ||  `|\_ . .|. .||]],
-    [[|| . _||   ||    || ||    ||   ||_ . || || . _||   ||    || ||   |\ `-_/| . ||]],
-    [[||_-' ||  .|/    || ||    \|.  || `-_|| ||_-' ||  .|/    || ||   | \  / |-_.||]],
-    [[||    ||_-'      || ||      `-_||    || ||    ||_-'      || ||   | \  / |  `||]],
-    [[||    `'         || ||         `'    || ||    `'         || ||   | \  / |   ||]],
-    [[||            .===' `===.         .==='.`===.         .===' /==. |  \/  |   ||]],
-    [[||         .=='   \_|-_ `===. .==='   _|_   `===. .===' _-|/   `==  \/  |   ||]],
-    [[||      .=='    _-'    `-_  `='    _-'   `-_    `='  _-'   `-_  /|  \/  |   ||]],
-    [[||   .=='    _-'          '-__\._-'         '-_./__-'         `' |. /|  |   ||]],
-    [[||.=='    _-'                                                     `' |  /==.||]],
-    [[=='    _-'                        N E O V I M                         \/   `==]],
-    [[\   _-'                                                                `-_   /]],
-    [[ `''                                                                      ``' ]],
-  },
-  opts = {
-    position = 'center',
-    hl = 'Type',
-    -- wrap = "overflow";
-  },
-}
-
-local footer = {
-  type = 'text',
-  val = '',
-  opts = {
-    position = 'center',
-    hl = 'Number',
-  },
-}
-
-local leader = '<leader>'
-
---- @param sc string
---- @param txt string
---- @param keybind string? optional
---- @param keybind_opts table? optional
-local function button(sc, txt, keybind, keybind_opts)
-  local sc_ = sc:gsub('%s', ''):gsub(leader, '<leader>')
-
-  local opts = {
-    position = 'center',
-    shortcut = sc,
-    cursor = 3,
-    width = 50,
-    align_shortcut = 'right',
-    hl_shortcut = 'Keyword',
-  }
-  if keybind then
-    keybind_opts = if_nil(keybind_opts, { noremap = true, silent = true, nowait = true })
-    opts.keymap = { 'n', sc_, keybind, keybind_opts }
-  end
-
-  local function on_press()
-    local key = vim.api.nvim_replace_termcodes(keybind or sc_ .. '<Ignore>', true, false, true)
-    vim.api.nvim_feedkeys(key, 't', false)
-  end
-
-  return {
-    type = 'button',
-    val = txt,
-    on_press = on_press,
-    opts = opts,
-  }
-end
-
-local buttons = {
-  type = 'group',
-  val = {
-    button('<leader>ff', '󰈞  Find file'),
-    button('<leader>fg', '󰈬  Find word'),
-    button('<leader>fh', '󰋖  Find help'),
-    button('<leader>fc', '  Find commit'),
-    button('<leader>fm', '  Find man page'),
-    button('e', '  New file', '<cmd>ene <CR>'),
-  },
-  opts = {
-    spacing = 1,
-  },
-}
-
-local section = {
-  terminal = default_terminal,
-  header = default_header,
-  buttons = buttons,
-  footer = footer,
-}
-
-M.layout = {
-    { type = 'padding', val = 2 },
-    section.header,
-    { type = 'padding', val = 2 },
-    section.buttons,
-    section.footer,
-  }
-  M.opts = {
-    margin = 5,
-  }
-
-return M
