+++ .config/nvim/lua/senyc/lazy.lua
@@ -11,4 +11,4 @@ if not vim.loop.fs_stat(lazypath) then
 end
 vim.opt.rtp:prepend(lazypath)
 
-require('lazy').setup('senyc.plugins')
+require 'lazy'.setup 'senyc.plugins'
