+++ .config/nvim/lua/senyc/plugins/context.lua
@@ -4,7 +4,7 @@ return {
   config = function()
     require 'treesitter-context'.setup {
       enable = true,        -- Enable this plugin (Can be enabled/disabled later via commands)
-      max_lines = 0,        -- How many lines the window should span. Values <= 0 mean no limit.
+      max_lines = 3,        -- How many lines the window should span. Values <= 0 mean no limit.
       min_window_height = 0, -- Minimum editor window height to enable context. Values <= 0 mean no limit.
       line_numbers = true,
       multiline_threshold = 20, -- Maximum number of lines to show for a single context
