+++ .config/nvim/lua/senyc/plugins/lualine.lua
@@ -2,7 +2,7 @@ return {
   'nvim-lualine/lualine.nvim',
   lazy = false,
   config = function()
-    require('lualine').setup {
+    require 'lualine'.setup {
       options = {
         icons_enabled = false,
         theme = 'catppuccin',
