+++ .config/nvim/lua/senyc/plugins/tailwind-sorter.lua
@@ -1,9 +1,9 @@
 return {
   'laytan/tailwind-sorter.nvim',
-  ft = { 'html', 'js', 'jsx', 'tsx', 'astro' },
   lazy = true,
+  ft = { 'html', 'js', 'jsx', 'tsx', 'astro' },
   config = function()
-    require('tailwind-sorter').setup({
+    require 'tailwind-sorter'.setup({
 
       on_save_enabled = true,                                                                                    -- If `true`, automatically enables on save sorting.
       on_save_pattern = { '*.html', '*.js', '*.jsx', '*.tsx', '*.astro' }, -- The file patterns to watch and sort.
