--- .config/nvim/lua/senyc/plugins/treesitter-playground.lua
@@ -1,6 +1,0 @@
-return {
-  'nvim-treesitter/playground',
-  dependencies = { 'nvim-treesitter/nvim-treesitter' },
-  lazy = true,
-  cmd = 'TSPlaygroundToggle',
-}
