+++ .config/nvim/lua/senyc/utils.lua
@@ -1,4 +1,3 @@
--- local lfs = require("lfs")
 local M = {}
 
 --- Returns the base git directory from CWD else return error code
@@ -19,30 +18,5 @@ function M.get_git_dir()
   end
   return 'No valid return file', nil
 end
-
---- Gets all of the files in a directory and writes them to a table
---- will not return files that are in the hidden git directory
----
--- ---@return table filetab
--- function M.get_all_files_in_dir(directory)
---   local files = {}
---   for file in lfs.dir(directory) do
---     -- Checks if string includes .git
---     -- the % is the escape character for the period
---     if file ~= "." and file ~= ".." and not file:match(".*%.git") then
---       local filepath = directory .. "/" .. file
---       local mode = lfs.attributes(filepath, "mode")
---       if mode == "file" then
---         table.insert(files, filepath)
---       elseif mode == "directory" then
---         local subFiles = M.get_all_files_in_dir(filepath)
---         for _, subFile in ipairs(subFiles) do
---           table.insert(files, subFile)
---         end
---       end
---     end
---   end
---   return files
--- end
 
 return M
